// src/@fake-db/types.ts
export type DataMemberType = {
  nik: string
  role: string
  attendances?: string | undefined
  statuses: string
  avatar?: string
  fullName: string
}